Move RewindableFileHandle to lib/bot.py
